requirements.txt
# # requirements.txt
# # Optimized for Python 3.10.11
# # Tooth Segmentation Active Learning System

# # Core ML Framework
# tensorflow==2.15.0
# keras==2.15.0

# # Web Interface with Sketchpad support
# gradio==4.44.1
# gradio-client==1.3.0

# # Data Processing
# numpy==1.26.4
# pillow==10.0.1
# opencv-python==4.8.1.78

# # Machine Learning Utilities
# scikit-learn==1.3.2

# # Database
# # (sqlite3 is built into Python - no need to install)

# # Additional Dependencies (auto-installed by above, but pinned for stability)
# matplotlib==3.8.2
# pandas==2.1.4
# scipy==1.11.4

# # Gradio Dependencies (explicitly pinned)
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.3
# jinja2==3.1.2
# python-multipart==0.0.9
# aiofiles==23.2.1
# httpx==0.25.2
# websockets==12.0
# markupsafe==2.1.3

# # Optional but recommended
# tqdm==4.66.1  # Progress bars during training

# requirements.txt - Gradio 5.24 Compatible
# For Python 3.10.11

# Core ML Framework
tensorflow==2.16.1
keras==3.0.5

# Web Interface - UPGRADED
gradio==5.24.0

# Data Processing
numpy==1.26.4
pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.2

# Utilities
matplotlib==3.8.2
tqdm==4.66.1

